Exec Dashboard Rearchitecture
After a couple days of effort, I was unable to get the exec dashboard
to function properly in our deployed application. As such I have
decided to rearchitect it, turning all the python code into app calls,
and converting the server code into NodeJS so that it in-line with our
existing architecture. The UI code was cleaned a bit, converted into
static files, and will be served from the admin page. Using the same
log-in authentication as the rest of the application. Note that this is
a cross repo commit, sharing the same title and description.