Clone nginx to simplify upgrading nginx in prod
The old nginx code is tightly intertwined in the common module, which
makes it difficult to get it into production. As a stopgap, this just
extracts out the code into a new module.

Note this uses a different AWS name so we can distinguish it from
the real production nginx server in ansible.

Note this puts the website up on www2 to test out it works before
we transition it to be the official load balancer.